Renamed test classes.
